@@ -17,9 +17,9 @@
     "@types/d3": "5.7.2",
     "@types/dagre": "0.7.43",
     "@types/node": "13.9.3",
-    "fuse-box": "3.7.1",
-    "terser": "4.6.10",
-    "ts-node": "8.6.2",
-    "typescript": "3.6.4"
+    "fuse-box": "4.0.0-next.438",
+    "terser": "4.7.0",
+    "ts-node": "8.10.2",
+    "typescript": "3.9.3"
   }
 }
